Remove usage of rewind
This sadly is not what I will be using this object for and I have no desire of "rewinding" once I loop through the object.
If this is something that is needed I will have stored the array anyways.
This allows us to use DBResult to grab the first row before actually entering an iterator pipe.